requirements.txt
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
pytest-django==4.5.2

